Main: Extract Rhizome creation.
The initRhizome function is doing a lot more than just preparing the data, as it is
replicating the example code from the rhizome library. It is also a standalone thing,
in that it creates a 'mock' Rhizome ready to go.

Create a new external mockRhizome function, that only creates a Rhizome with information
in it, and call it in main.