WIP: Stub the RhizomeManager.
We're starting to see the rhizome parameter everywhere. Because
the Rhizome is essentially a singleton, we should be able to just define
a simple factory method to return the singular instance. That way there
is no longer a need to pass it down through a widget hierarchy, etc.

Stub a RhizomeManager that shall provide a static method for obtaining
the Rhizome anywhere it is needed.